skills/creator-crm/requirements.txt
pydantic>=2.0.0
requests>=2.31.0
playwright>=1.40.0
curl_cffi>=0.5.0
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
